Crate drone_nrf_map

Nordic Semi nRFx peripheral mappings for Drone, an Embedded Operating System.

This crate uses CMSIS-SVD files provided by Nordic Semiconductor to automatically generate Drone register and interrupt bindings. However only the corresponding Product Specification is the single source of truth. A difference between this crate bindings and the Product Specification is considered a bug. Fixing such a bug is not a breaking change.

This crate re-exports the contents of drone_cortexm::map module and is a drop-in replacement for it.

Supported Devices

nrf_mcu	Core name	Product specification	Available features

nrf52810	ARM® Cortex®-M4F r0p1	PS v1.3

nrf52811	ARM® Cortex®-M4F r0p1	PS v1.0

nrf52832	ARM® Cortex®-M4F r0p1	PS v1.4

nrf52840	ARM® Cortex®-M4F r0p1	PS v1.1

nrf9160	ARM® Cortex®-M33F r0p2	PS v1.1	uarte

nrf_mcu config flag should be set at the application level according to this table.

Documentation

Drone Book

API documentation

The API documentation intentionally skips auto-generated reg and thr bindings. Otherwise it would use several gigabytes of space and would be very slow to render in a browser. One should refer to the Product Specification instead. And to get an idea of what the API looks like on the Drone side, look at the drone_cortexm::map module documentation.

Usage

Add the crate to your Cargo.toml dependencies:

[dependencies]
drone-nrf-map = { version = "0.14.0", features = [...] }

Add or extend std feature as follows:

[features]
std = ["drone-nrf-map/std"]

Modules

periph
nRFx peripheral mappings.

reg
nRF register mappings.

thr
nRF interrupt mappings.

Macros

nrf_reg_tokens
Defines an index of nrf52 register tokens.